Implement UpdateContext
