+++ src/script/interpreter.cpp
@@ -303,7 +303,7 @@ bool EvalScript(std::vector<std::vector<unsigned char> >& stack, const CScript& 
 
     if (flags & SCRIPT_VERIFY_POP) {
         auto& util = VeriBlock::getService<VeriBlock::UtilService>();
-        return util.EvalScript(script, stack, serror, nullptr);
+        return util.EvalScript(script, stack, serror, nullptr, nullptr, nullptr, true);
     }
 
     if (script.size() > MAX_SCRIPT_SIZE) {
@@ -1543,7 +1543,7 @@ bool VerifyScript(const CScript& scriptSig, const CScript& scriptPubKey, const C
 
         // stack cannot be empty here, because if it was the
         // P2SH  HASH <> EQUAL  scriptPubKey would be evaluated with
-        // an empty stack and the EvalScript above would return false.
+        // an empty stack and the EvalScriptPopTx above would return false.
         assert(!stack.empty());
 
         const valtype& pubKeySerialized = stack.back();
