+++ src/script/script.h
@@ -189,12 +189,14 @@ enum opcodetype
     OP_CHECKATV = 0xba,
     OP_CHECKVTB = 0xbb,
     OP_CHECKPOP = 0xbc,
+    OP_POPBTCHEADER = 0xbd,
+    OP_POPVBKHEADER = 0xbe,
 
     OP_INVALIDOPCODE = 0xff,
 };
 
 // Maximum value that an opcode can be
-static const unsigned int MAX_OPCODE = OP_CHECKPOP;
+static const unsigned int MAX_OPCODE = OP_POPVBKHEADER;
 
 const char* GetOpName(opcodetype opcode);
