--- src/vbk/entity/publications.hpp
@@ -1,18 +1,0 @@
-#ifndef INTEGRATION_REFERENCE_BTC_PUBLICATIONS_HPP
-#define INTEGRATION_REFERENCE_BTC_PUBLICATIONS_HPP
-
-#include <stdint.h>
-#include <vector>
-
-namespace VeriBlock {
-
-using ByteArray = std::vector<uint8_t>;
-
-struct Publications {
-    ByteArray atv;
-    std::vector<ByteArray> vtbs;
-};
-
-} // namespace VeriBlock
-
-#endif //INTEGRATION_REFERENCE_BTC_PUBLICATIONS_HPP
