+++ src/vbk/test/integration/utils.hpp
@@ -16,4 +16,5 @@ void setUpGenesisBlocks();
 
 VeriBlock::AltPublication generateSignedAltPublication(const CBlock& endorsedBlock, const uint32_t& identifier, const CScript& payoutInfo);
 } // namespace VeriBlockTest
+
 #endif
